% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{encoding_tag}
\alias{encoding_tag}
\title{タグのエンコード}
\usage{
encoding_tag(x)
}
\arguments{
\item{x}{character タグ sample_tag,テスト,test_tag または c("sample_tag","テスト","test_tag")}
}
\value{
character "sample_tag,テスト,test_tag"をbase64でエンコードした結果(サンプルは"c2FtcGxlX3RhZyzjg4bjgrnjg4gsdGVzdF90YWc=")
}
\description{
headerにタグを含ませる場合、utf-8で送れないので、base64でエンコードする
}
\examples{
encoding_tag("sample_tag,テスト,test_tag")
encoding_tag(c("sample_ t a  g ","テスト "," test_tag"))
}
