% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{encoding_title}
\alias{encoding_title}
\title{タイトルのエンコード}
\usage{
encoding_title(x)
}
\arguments{
\item{x}{character タイトル}
}
\value{
character タイトルをbase64でエンコードした結果
}
\description{
headerにタイトルを含ませる場合、utf-8で送れないので、base64でエンコードする
}
\examples{
encoding_title("test_title")
encoding_title("サンプルタイトル")
}
