% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_addr}
\alias{expand_addr}
\title{Expand postal addresses}
\usage{
expand_addr(addresses)
}
\arguments{
\item{addresses}{a character vector of addresses to normalise}
}
\value{
a list of character vectors of expanded addresses.
}
\description{
\code{expand_addr} takes street
addresses and expands them to one or more normalised strings
within the context of a specific
country or language. It currently recognises 60 different locales.
}
\examples{
# expand an English address!
\dontrun{
expand_addr("fourty seven love lane pinner")
# [1] "47 love lane pinner"
}
# Normalise a French one
\dontrun{
expand_addr("Quatre-vignt-douze Ave des Champs-Élysées")
# "92 avenue des champs-elysees"
}
}
\seealso{
\code{\link{parse_addr}} for parsing addresses.
}
