% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalise_addr}
\alias{normalise_addr}
\title{Normalise postal addresses}
\usage{
normalise_addr(addresses)
}
\arguments{
\item{addresses}{a character vector of addresses to normalise}
}
\value{
a character vector of normalised addresses.
}
\description{
\code{normalise_addr} takes street
addresses and normalises them within the context of a specific
country or language. It currently recognises 60 different locales.
}
\examples{
# Normalise an English address!
normalise_addr("fourty seven love lane pinner")
# [1] "47 love lane pinner"

# Normalise a French one
normalise_addr("Quatre-vignt-douze Ave des Champs-Élysées")
# "92 avenue des champs-elysees"

}
\seealso{
\code{\link{parse_addr}} for parsing addresses.
}

