% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_addr}
\alias{parse_addr}
\title{Parse street addresses}
\usage{
parse_addr(addresses)
}
\arguments{
\item{addresses}{a character vector of addresses to parse.}
}
\value{
a data.frame of 20 columns; \code{house}, \code{category},
\code{near}, \code{house_number}, \code{road}, \code{unit},
\code{level}, \code{staircase}, \code{entrance}, \code{po_box},
\code{suburb}, \code{city_district}, \code{city},
\code{state_district}, \code{state}, \code{postal_code},
\code{country_region}, \code{country}, \code{world_region}. 
Values not found in the address are represented
with \code{NA}s
}
\description{
\code{parse_addr} parses street addresses into
their component parts, producing the addresses' house name,
number, road and suburb, city district and city, state district
and state, postal code, and country.
}
\examples{
\dontrun{
str(parse_addr("781 Franklin Ave Crown Heights Brooklyn NYC NY 11216 USA"))
}
# 'data.frame':	1 obs. of  10 variables:
#   $ house          : chr NA
#   $ category       : chr NA
#   $ near           : chr NA
#   $ house_number   : chr "781"
#   $ road           : chr "franklin ave"
#   $ unit           : chr NA
#   $ level          : chr NA
#   $ staircase      : chr NA
#   $ entrance       : chr NA
#   $ po_box         : chr NA
#   $ suburb         : chr "crown heights"
#   $ city_district  : chr "brooklyn"
#   $ city           : chr "nyc"
#   $ state_district : chr NA
#   $ state          : chr "ny"
#   $ postal_code    : chr NA
#   $ country_region : chr NA
#   $ country        : chr "usa"
#   $ world_region   : chr NA

}
\seealso{
\code{\link{normalise_addr}} for normalising addresses.
}
