% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_html.R
\name{post_html}
\alias{post_html}
\title{htmlレポートを転送する}
\usage{
post_html(path)
}
\arguments{
\item{path}{character レポートへのpath}
}
\value{
numeric textのポジション
}
\description{
htmlレポートを転送する
}
\examples{
post_html("aaaa")
}
