% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_html.R
\name{post_html_from_Rmd}
\alias{post_html_from_Rmd}
\title{htmlレポートをPOSTする}
\usage{
post_html_from_Rmd(Rmd_file, interactive = NULL)
}
\arguments{
\item{Rmd_file}{character レポートへのpath}
}
\value{
\code{post_html_from_Rmd()} returns the URL of the html page.
}
\description{
Rmdファイルから、生成されたhtmlをPOSTする。htmlレポートがない場合は、生成を聞く。
}
\examples{
post_html_from_Rmd("README.Rmd")
}
