% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_html.R
\name{post_html_from_Rmd_addin}
\alias{post_html_from_Rmd_addin}
\title{htmlレポート転送用アドイン関数}
\usage{
post_html_from_Rmd_addin()
}
\value{
\code{post_html_from_Rmd()} returns the URL of the html page.
}
\description{
RstudioのAddinsで関数の起動ができるようにする。開いているRmdファイルを対象にする
}
