% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteR-package.R
\docType{package}
\name{posteR-package}
\alias{posteR}
\alias{posteR-package}
\title{posteR: What the Package Does (One Line, Title Case)}
\description{
Provides utilities for posting html files to some API / server.
}
\author{
\strong{Maintainer}: Taiki Kubota \email{first.last@example.com}

}
\keyword{internal}
