% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poster.R
\docType{package}
\name{poster}
\alias{poster}
\alias{poster-package}
\title{Tools for handling URLs}
\description{
Normalise street addresses, or parse them, extracting house
  number, street name, post code and various other elements of a street
  address. Requires (and is built around) the libpostal library
  ('https://github.com/openvenues/libpostal').
}
\seealso{
the package \code{vignette("Address Parsing with Poster", package = "poster")}.
}
