% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{house}
\alias{house}
\alias{house_number}
\alias{road}
\alias{suburb}
\alias{city_district}
\alias{city}
\alias{state_district}
\alias{state}
\alias{postal_code}
\alias{country}
\title{Get an address's elements}
\usage{
house(x)

house_number(x)

road(x)

suburb(x)

city_district(x)

city(x)

state_district(x)

state(x)

postal_code(x)

country(x)
}
\arguments{
\item{x}{an address, or vector of addresses}
}
\description{
as in the lubridate package, individual components of an address
can be extracted using the relevant function call - see the
examples.
}
\examples{
#Get a component
\dontrun{
example_address <- "The Deaconage, 781 Franklin Ave Crown Heights Brooklyn NYC NY 11216 USA"
house(example_address)
}
}
