% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_addr}
\alias{parse_addr}
\title{Parse street addresses}
\usage{
parse_addr(addresses)
}
\arguments{
\item{addresses}{a character vector of addresses to parse.}
}
\value{
a data.frame of 10 columns; \code{house}, \code{house_number},
\code{road}, \code{suburb}, \code{city_district}, \code{city},
\code{state_district}, \code{state}, \code{postal_code},
\code{country}. Values not found in the address are represented
with \code{NA}s
}
\description{
\code{parse_addr} parses street addresses into
their component parts, producing the addresses' house name,
number, road and suburb, city district and city, state district
and state, postal code, and country.
}
\examples{
parse_addr("781 Franklin Ave Crown Heights Brooklyn NYC NY 11216 USA")

}

