% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_html.R, R/posteR-package.R
\docType{package}
\name{posteR-package}
\alias{posteR}
\alias{posteR-package}
\title{Example package to say hello}
\description{
Provides utilities for posting html files to some API / server.

Provides utilities for posting html files to some API / server.
}
\author{
\strong{Maintainer}: Taiki Kubota \email{first.last@example.com}


\strong{Maintainer}: Taiki Kubota \email{first.last@example.com}

}
\keyword{internal}
